Trial to add search box
When I try to add the autocomplete search box, then something strange happens. First load it shows the map with the location of the taps but search box doesn't work.
Upon reload the search box works but the pins are not there anymore
I can't solve it, I think it is something with the callbacks and how I need to structure the js. 
I tried a lot of options, but I can't get it work. 
Alexa, could you pls take a look at the code and give me a helping hand?